`GatwayManager.id: Int` -> `UInt`

+++ Sources/DiscordGateway/BotGatewayManager.swift
@@ -15,7 +15,7 @@ public actor BotGatewayManager: GatewayManager {
     /// Max frame size we accept to receive through the websocket connection.
     nonisolated let maxFrameSize: Int
     /// Generator of `BotGatewayManager` ids.
-    static let idGenerator = ManagedAtomic(0)
+    static let idGenerator = ManagedAtomic(UInt(0))
     /// This gateway manager's identifier.
     public nonisolated let id = BotGatewayManager.idGenerator
         .wrappingIncrementThenLoad(ordering: .relaxed)
